reflectorwatcher using store, and moving resyncrunnable
